Activate mcp server
If we simply activate it in the same container as the main ragflow
processes, then the mcp process will prevent auto-restarting the main
ragflow app when elasticsearch is not available for 2 mins. Then, the
application stays permanently down. So the quickest solution is to put
the mcp server in a second container in the same pod.

Better would be to put it in a separate deployment.