Re-add mysql
Turns out the MySQL of the PBS Betriebsplattform is actually a MariaDB,
which is fine.
But it also turns out that RAGFlow is not 100% compatible with MariaDB.
The progress of parsing files cannot be reported, because RAGFlow tries
to acquire a lock, and MariaDB and MySQL work differently with regard to
the locking functions.
So for now, we host a MySQL pod again. Maybe we can switch to Postgres
later.